requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch==2.0.1+cu117
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
sentence-transformers>=2.2.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
python-dotenv>=0.19.0
ipykernel>=6.0.0
gradio==3.50.2
boto3==1.28.15
awscli==1.29.15
pydantic==1.10.9
onnxruntime==1.14.1